[ADD] clearbutton을 textfield extension으로 빼기(#14)
